+++ influxdb_client/configuration.py
@@ -231,7 +231,7 @@ class Configuration(six.with_metaclass(TypeWithDefault, object)):
                "OS: {env}\n"\
                "Python Version: {pyversion}\n"\
                "Version of the API: 0.1.0\n"\
-               "SDK Package Version: 1.5.0dev".\
+               "SDK Package Version: 1.5.0".\
                format(env=sys.platform, pyversion=sys.version)
 
     def update_request_header_params(self, path: str, params: dict):