+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Bugs
 1. [#61](https://github.com/influxdata/influxdb-client-python/issues/61): Correctly parse CSV where multiple results include multiple tables
+1. [#66](https://github.com/influxdata/influxdb-client-python/issues/66): Correctly close connection pool manager at exit
 
 ## 1.4.0 [2020-02-14]